+ Added _fragments file with 'placeholder' fragment. + Friends requests now appearing in accordion view. + Updated types in friend_requests table.

+++ src/main/java/com/makersacademy/acebook/repository/FriendRequestRepository.java
@@ -7,6 +7,6 @@ import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 public interface FriendRequestRepository extends CrudRepository<FriendRequest, Long> {
-    public ArrayList<FriendRequest> findFriendRequestByfromUserIdAndStatus(int fromUserId,String status);
-    public ArrayList<FriendRequest> findFriendRequestBytoUserIdAndStatus(int fromUserId,String status);
+    public ArrayList<FriendRequest> findFriendRequestByfromUserIdAndStatus(long fromUserId,String status);
+    public ArrayList<FriendRequest> findFriendRequestBytoUserIdAndStatus(long fromUserId,String status);
 }
